refactor: centralize Slurm scalar types
